Almost done

+++ controllers/cards.js
@@ -1,89 +1,77 @@
-const mongoose = require('mongoose');
 const Card = require('../models/card');
 const {
-  NOT_FOUND_ERROR,
-  VALIDATION_ERROR,
-  SERVER_ERROR,
+  STATUS_CODES,
   ERROR_MESSAGE,
 } = require('../utils/constants');
 
-const createCards = (req, res) => {
+const NotFoundErr = require('../errors/NotFoundErr');
+const ForbbidenErr = require('../errors/ForbiddenErr');
+const BadRequestErr = require('../errors/BadRequestErr');
+
+const createCards = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
 
   Card.create({ name, link, owner })
     .then((card) => {
-      res.status(201).send({ data: card });
+      res.status(STATUS_CODES.SUCCESS).send({ data: card });
     })
     .catch((err) => {
-      if (err instanceof mongoose.Error.ValidationError) {
-        return res.status(VALIDATION_ERROR).send({ message: ERROR_MESSAGE.GET_CARDS_ERROR });
+      if (err.name === 'ValidationError') {
+        return next(new BadRequestErr(ERROR_MESSAGE.GET_CARDS_ERROR));
       }
-      return res.status(SERVER_ERROR).send({ message: ERROR_MESSAGE.SERVER_ERROR });
+      return next(err);
     });
 };
 
-const getCards = (req, res) => {
+const getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => {
       res.send({ data: cards });
     })
-    .catch(() => {
-      res.status(SERVER_ERROR).send({ message: ERROR_MESSAGE.SERVER_ERROR });
-    });
+    .catch(next);
 };
 
-const deleteCardById = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+const deleteCardById = (req, res, next) => {
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
-        return res.status(NOT_FOUND_ERROR).send({ message: ERROR_MESSAGE.DELETE_CARDSID_ERROR });
+        throw new NotFoundErr(ERROR_MESSAGE.DELETE_CARDSID_ERROR);
       }
-      return res.send({ data: card });
+      if (card.owner.toString() !== req.user._id) {
+        throw new ForbbidenErr(ERROR_MESSAGE.REFUSAL_TO_DELETE);
+      }
+      return Card.findByIdAndRemove(req.params.cardId)
+        .then((removedCard) => res.send(removedCard));
     })
-    .catch((err) => {
-      if (err instanceof mongoose.Error.CastError) {
-        return res.status(VALIDATION_ERROR).send({ message: ERROR_MESSAGE.LIKE_CARDID_VALIDAT_ER });
-      }
-      return res.status(SERVER_ERROR).send({ message: ERROR_MESSAGE.SERVER_ERROR });
-    });
+    .catch(next);
 };
 
-const likeCard = (req, res) => Card.findByIdAndUpdate(
+const likeCard = (req, res, next) => Card.findByIdAndUpdate(
   req.params.cardId,
   { $addToSet: { likes: req.user._id } },
   { new: true },
 )
   .then((card) => {
     if (!card) {
-      return res.status(NOT_FOUND_ERROR).send({ message: ERROR_MESSAGE.DELETE_CARDSID_ERROR });
+      throw new NotFoundErr(ERROR_MESSAGE.DELETE_CARDSID_ERROR);
     }
     return res.send({ data: card });
   })
-  .catch((err) => {
-    if (err instanceof mongoose.Error.CastError) {
-      return res.status(VALIDATION_ERROR).send({ message: ERROR_MESSAGE.LIKE_CARDID_VALIDAT_ER });
-    }
-    return res.status(SERVER_ERROR).send({ message: ERROR_MESSAGE.SERVER_ERROR });
-  });
+  .catch(next);
 
-const dislikeCard = (req, res) => Card.findByIdAndUpdate(
+const dislikeCard = (req, res, next) => Card.findByIdAndUpdate(
   req.params.cardId,
   { $pull: { likes: req.user._id } },
   { new: true },
 )
   .then((card) => {
     if (!card) {
-      return res.status(NOT_FOUND_ERROR).send({ message: ERROR_MESSAGE.DELETE_CARDSID_ERROR });
+      throw new NotFoundErr(ERROR_MESSAGE.DELETE_CARDSID_ERROR);
     }
     return res.send({ data: card });
   })
-  .catch((err) => {
-    if (err instanceof mongoose.Error.CastError) {
-      return res.status(VALIDATION_ERROR).send({ message: ERROR_MESSAGE.LIKE_CARDID_VALIDAT_ER });
-    }
-    return res.status(SERVER_ERROR).send({ message: ERROR_MESSAGE.SERVER_ERROR });
-  });
+  .catch(next);
 
 module.exports = {
   createCards,
